Backend/tests/requirements.txt
# ============================================
# Production Dependencies
# ============================================

# Web Framework
fastapi==0.128.0
starlette==0.50.0  # FastAPI dependency
pydantic==2.12.5
pydantic-core==2.41.5

# ASGI Server
uvicorn==0.40.0
h11==0.16.0  # uvicorn dependency

# Database
psycopg2-binary==2.9.11

# Configuration
python-dotenv==1.2.1

# HTTP Client (for external API calls if needed)
httpx==0.28.1
httpcore==1.0.9

# Utilities
python-multipart==0.0.21  # For file uploads
typing-extensions==4.15.0

# ============================================
# Optional: Data Processing Dependencies
# (Include only if used in production)
# ============================================
pandas==2.2.3
numpy==2.2.4
openpyxl==3.1.5  # Excel file support

# PDF Processing (for extractos)
pdfplumber==0.11.8
pdfminer.six==20251107
pypdfium2==5.2.0

# String matching (for clasificación automática)
thefuzz==0.22.1
python-Levenshtein==0.27.3
Levenshtein==0.27.3
RapidFuzz==3.14.3
